+++ lib/src/controls/inputs/buttons/outlined_button.dart
@@ -14,6 +14,8 @@ class OutlinedButton extends BaseButton {
     required Widget child,
     required VoidCallback? onPressed,
     VoidCallback? onLongPress,
+    VoidCallback? onTapDown,
+    VoidCallback? onTapUp,
     FocusNode? focusNode,
     bool autofocus = false,
     ButtonStyle? style,
@@ -25,6 +27,8 @@ class OutlinedButton extends BaseButton {
           autofocus: autofocus,
           onLongPress: onLongPress,
           onPressed: onPressed,
+          onTapDown: onTapDown,
+          onTapUp: onTapUp,
           style: style,
           focusable: focusable,
         );